requirements.txt
# requirements.txt - Dependencies for MASB-Alt

# Core dependencies
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
# sqlite3 is part of Python standard library

# LLM API clients
openai==1.3.0
anthropic==0.7.0
cohere==4.37.0
aiohttp==3.9.0
# asyncio is part of Python standard library

# Evaluation and analysis
scikit-learn==1.3.0
scipy==1.11.1
tqdm==4.66.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
streamlit==1.28.0
kaleido==0.2.1  # For plotly image export

# GUI dependencies (for human evaluation tool)
tkinter==0.1.0  # Usually comes with Python

# Web framework (optional, for API endpoints)
fastapi==0.104.0
uvicorn==0.24.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0

# Logging and monitoring
loguru==0.7.2
sentry-sdk==1.38.0  # Optional: for error tracking

# Data validation
pydantic==2.5.0
jsonschema==4.20.0